# openapi.yaml
"openapi": "3.0.3"
"info":
  "title": "EX2"
  "version": "1.0.0"
  "contact": {}
"servers":
  -
    "url": "localhost"
"paths":
...
  "/admin/course":
    "post":
      "tags":
        - "Admin"
        - "Course"
      "summary": "Add new course"
      "description": "Add new course"
      "operationId": "addNewCourse"
      "parameters":
        -
          "in": "header"
          "name": "Authorization"
          "schema":
            "type": "string"
          "required": true
      "requestBody":
        "content":
          "application/json":
            "schema":
              "type": "object"
              "properties":
                "corequisites":
                  "type": "array"
                  "items":
                    "type": "string"
                    "example": "3"
                  "example":
                    - "3"
                    - "4"
                "name":
                  "type": "string"
                  "example": "Math"
                "prerequisites":
                  "type": "array"
                  "items":
                    "type": "string"
                    "example": "1"
                  "example":
                    - "1"
                    - "2"
                "unitCount":
                  "type": "number"
                  "example": 3
            "example":
              "corequisites":
                - "3"
                - "4"
              "name": "Math"
              "prerequisites":
                - "1"
                - "2"
              "unitCount": 3
      "responses":
        "200":
          "description": ""
  "/admin/course/{id}":
    "get":
      "tags":
        - "Admin"
        - "Course"
      "summary": "Fetch course by id"
      "description": "Fetch course by id"
      "operationId": "fetchCourseById"
      "parameters":
        -
          "in": "header"
          "name": "Authorization"
          "schema":
            "type": "string"
          "required": true
      "responses":
        "200":
          "description": ""
    "put":
      "tags":
        - "Admin"
        - "Course"
      "summary": "update course"
      "description": "update course"
      "operationId": "updateCourse"
      "parameters":
        -
          "in": "header"
          "name": "Authorization"
          "schema":
            "type": "string"
          "required": true
      "requestBody":
        "content":
          "application/json":
            "schema":
              "type": "object"
              "properties":
                "corequisites":
                  "type": "array"
                  "items":
                    "type": "string"
                    "example": "3"
                  "example":
                    - "3"
                    - "4"
                "name":
                  "type": "string"
                  "example": "Math new"
                "prerequisites":
                  "type": "array"
                  "items":
                    "type": "string"
                    "example": "1"
                  "example":
                    - "1"
                    - "2"
                "unitCount":
                  "type": "number"
                  "example": 3
            "example":
              "corequisites":
                - "3"
                - "4"
              "name": "Math new"
              "prerequisites":
                - "1"
                - "2"
              "unitCount": 3
      "responses":
        "200":
          "description": ""
    "delete":
      "tags":
        - "Admin"
        - "Course"
      "summary": "Delete course"
      "description": "Delete course"
      "operationId": "deleteCourse"
      "parameters":
        -
          "in": "header"
          "name": "Authorization"
          "schema":
            "type": "string"
          "required": true
      "responses":
        "200":
          "description": ""
    "parameters":
      -
        "name": "id"
        "in": "path"
        "required": true
        "schema":
          "type": "string"
...